commit jar package of JavaFX app
